Allow multiple subscriptions to useRegisterWebsocketListener (#423)
* Refactor usage of useRegisterWebsocketListener
* Format Readme and add info about useRegisterWebsocketListener
* Create useGetWebsocketEvent hook
