+++ src/hooks/account/index.ts
@@ -3,3 +3,4 @@ export * from './useGetAccount';
 export * from './useGetLoginInfo';
 export * from './useGetIsLoggedIn';
 export * from './useGetAccountProvider';
+export * from './useGetWebsocketEvent';
