+++ src/hooks/account/useGetWebsocketEvent.ts
@@ -1,0 +1,6 @@
+import { useSelector } from 'reduxStore/DappProviderContext';
+import { websocketEventSelector } from 'reduxStore/selectors';
+
+export const useGetWebsocketEvent = () => {
+  return useSelector(websocketEventSelector);
+};
