+++ src/reduxStore/slices/accountInfoSlice.ts
@@ -31,6 +31,10 @@ export interface AccountInfoSliceType {
   ledgerAccount: LedgerAccountType | null;
   walletConnectAccount: string | null;
   isAccountLoading: boolean;
+  websocketEvent: {
+    timestamp: number;
+    message: string;
+  } | null;
   accountLoadingError: string | null;
 }
 
@@ -45,6 +49,7 @@ export const emptyAccount: AccountType = {
 
 const initialState: AccountInfoSliceType = {
   address: '',
+  websocketEvent: null,
   accounts: { '': emptyAccount },
   ledgerAccount: null,
   publicKey: '',
@@ -124,6 +129,15 @@ export const accountInfoSlice = createSlice({
     ) => {
       state.accountLoadingError = action.payload;
       state.isAccountLoading = false;
+    },
+    setWebsocketEvent: (
+      state: AccountInfoSliceType,
+      action: PayloadAction<string>
+    ) => {
+      state.websocketEvent = {
+        timestamp: Date.now(),
+        message: action.payload
+      };
     }
   },
   extraReducers: (builder) => {
@@ -172,7 +186,8 @@ export const {
   updateLedgerAccount,
   setWalletConnectAccount,
   setIsAccountLoading,
-  setAccountLoadingError
+  setAccountLoadingError,
+  setWebsocketEvent
 } = accountInfoSlice.actions;
 
 export default accountInfoSlice.reducer;
